rtl: replace sim rom with a dual-ported ram.
This should be enough to start implementing some tests that write to
memory, but it's just a placeholder until we get a real memory (SDRAM),
but this means we don't have to worry about wait states yet.

This isn't really part of the CPU core so move it to the top level.