rtl/decode: handle branch/call to ra.
For branch/call, when bit 25 is set we branch to the contents of ra.  We
can optimize some of this later by removing the explicit check for
OPCODE_RET if we have ret always have bit 25 set (RET without bit 25
doesn't make any sense, and we could also include $lr as the RET
operand).