debugger/tests: add a testpoint infrastructure.
testpoints are a mechanism to bind addresses to tests that can be easily
queried by the debugger.  For example we can mark execution of an
instruction as being a successful or failing test, or a custom testpoint
to allow a test script to query the state of the simulation and then
make a success/failure decision.