rtl/decode: convert to microcode based decoding.
Rather than using logic which doesn't scale well, generate the decode
signals from a ROM which is in turn generated from the instruction
definition and a python script.  This makes it easier to add new
instructions and avoiding bugs.

On a device that doesn't have asynchronous memories this may not result
in a saving in gates, but it's an RTL code saving at least.