oldland: implement debugger breakpoints.
The bkp instruction is used for stopping execution.  At the moment this
is only valid when the debugger attached, eventually the bkp instruction
should raise an exception when the debugger is not connected.

One hitting a breakpoint the debugger needs to set the PC to the
instruction following the breakpoint to resume execution.