sim: indicate which items are cacheable.
For implementing the data cache we need to know what peripherals are
cacheable or not.  For instruction fetches everything is cacheable, but
for data accesses only ram/rom is cacheable.  Rather than hard code that
into the load/store helpers, provide a helper for getting the
cacheability of an address.

Note that later once an MMU is implemented this will still be the
canonical source - a page marked as cacheable won't be cacheable if the
peripheral is not really cacheable.