oldland_cache: rework cache to allow writes.
Rework the cache so that it can be used as a data cache and hook it up
to the CPU.  This implements a simple data cache but does not yet track
dirty lines, only valid.  This means that when evicting a line due to a
conflict may cause the line to be written back even if not modified.